put sizes back to uint32_t.

+++ src/hotspot/share/utilities/elfFile.cpp
@@ -1451,7 +1451,7 @@ bool DwarfFile::LineNumberProgram::apply_standard_opcode(const uint8_t opcode) {
         // Must be at most 4 bytes because the index register is only 4 bytes wide.
         return false;
       }
-      uint8_t operation_advance = checked_cast<uint8_t>(adv);
+      uint32_t operation_advance = checked_cast<uint32_t>(adv);
       _state->add_to_address_register(operation_advance, _header);
       if (_state->_dwarf_version == 4) {
         _state->set_index_register(operation_advance, _header);
@@ -1708,7 +1708,7 @@ void DwarfFile::LineNumberProgram::LineNumberProgramState::reset_fields() {
 }
 
 // Defined in section 6.2.5.1 of the DWARF 4 spec.
-void DwarfFile::LineNumberProgram::LineNumberProgramState::add_to_address_register(const uint8_t operation_advance,
+void DwarfFile::LineNumberProgram::LineNumberProgramState::add_to_address_register(const uint32_t operation_advance,
                                                                                    const LineNumberProgramHeader& header) {
   if (_dwarf_version == 2 || _dwarf_version == 3) {
     _address += (uintptr_t)(operation_advance * header._minimum_instruction_length);
@@ -1719,7 +1719,7 @@ void DwarfFile::LineNumberProgram::LineNumberProgramState::add_to_address_regist
 }
 
 // Defined in section 6.2.5.1 of the DWARF 4 spec.
-void DwarfFile::LineNumberProgram::LineNumberProgramState::set_index_register(const uint8_t operation_advance,
+void DwarfFile::LineNumberProgram::LineNumberProgramState::set_index_register(const uint32_t operation_advance,
                                                                               const LineNumberProgramHeader& header) {
   _op_index = (_op_index + operation_advance) % header._maximum_operations_per_instruction;
 }
